+++ src/main/java/Grammar/Expression/Expression.java
@@ -6,7 +6,7 @@ import GamePlay.Territory;
 import java.util.Map;
 
 public interface Expression {
-    double eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError;
+    long eval(Map<String, Integer> bindings, Player p, Territory t) throws EvalError;
     void prettyPrint(StringBuilder s);
 }
